Raspberry/requirements.txt
aiocoap==0.4.7
asyncio==3.4.3
pyserial==3.5
python-decouple==3.8
